Move the tests in the `rclrs_tests` crate into `rclrs`.
