+++ rclrs/src/clock.rs
@@ -195,11 +195,10 @@ unsafe impl Send for rcl_clock_t {}
 mod tests {
     use super::*;
 
-    fn assert_send<T: Send>() {}
-    fn assert_sync<T: Sync>() {}
-
     #[test]
-    fn clock_is_send_and_sync() {
+    fn traits() {
+        use crate::test_helpers::*;
+
         assert_send::<Clock>();
         assert_sync::<Clock>();
     }
