+++ rclrs/src/wait.rs
@@ -404,11 +404,10 @@ impl WaitSet {
 mod tests {
     use super::*;
 
-    fn assert_send<T: Send>() {}
-    fn assert_sync<T: Sync>() {}
-
     #[test]
-    fn wait_set_is_send_and_sync() {
+    fn traits() {
+        use crate::test_helpers::*;
+
         assert_send::<WaitSet>();
         assert_sync::<WaitSet>();
     }
