--- rclrs_tests/package.xml
@@ -1,19 +1,0 @@
-<?xml version="2.0"?>
-<?xml-model
-   href="http://download.ros.org/schema/package_format3.xsd"
-   schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>rclrs_tests</name>
-  <version>0.4.1</version>
-  <description>Package containing tests for rclrs that make use of msgs and other dependencies</description>
-  <maintainer email="christopherjreid@outlook.com">Chris Reid</maintainer>
-  <license>Apache License 2.0</license>
-
-  <depend>test_msgs</depend>
-  <depend>rclrs</depend>
-  <depend>rosidl_runtime_rs</depend>
-
-  <export>
-    <build_type>ament_cargo</build_type>
-  </export>
-</package>
