--- rclrs_tests/src/lib.rs
@@ -1,5 +1,0 @@
-#![cfg(test)]
-
-mod client_service_tests;
-mod graph_tests;
-mod pub_sub_tests;
